move request-map-style into its own side effect component
